+++ bindings/python/System/src/LinearTask.cpp
@@ -1,15 +1,15 @@
 /**
- * @file VariablesHandler.cpp
- * @authors Giulio Romualdi
+ * @file LinearTask.cpp
+ * @authors Paolo Maria Viceconte
  * @copyright 2021 Istituto Italiano di Tecnologia (IIT). This software may be modified and
  * distributed under the terms of the GNU Lesser General Public License v2.1 or any later version.
  */
 
 #include <pybind11/pybind11.h>
+#include <pybind11/stl.h>
 
+#include <BipedalLocomotion/System/LinearTask.h>
 #include <BipedalLocomotion/bindings/System/LinearTask.h>
-#include <BipedalLocomotion/bindings/System/Module.h>
-#include <BipedalLocomotion/bindings/System/VariablesHandler.h>
 
 namespace BipedalLocomotion
 {
@@ -17,13 +17,16 @@ namespace bindings
 {
 namespace System
 {
-void CreateModule(pybind11::module& module)
-{
-    module.doc() = "System module";
 
-    CreateVariablesHandler(module);
-    CreateLinearTask(module);
+void CreateLinearTask(pybind11::module& module)
+{
+    namespace py = ::pybind11;
+
+    using namespace BipedalLocomotion::System;
+
+    py::class_<LinearTask, std::shared_ptr<LinearTask>>(module, "LinearTask");
 }
+
 } // namespace System
 } // namespace bindings
 } // namespace BipedalLocomotion